last commit

+++ blog-backend/src/routes/posts.js
@@ -1,5 +1,6 @@
 import PostModel from "../db/models/PostModel.js"
 import CommentsModel from "../db/models/CommentsModel.js"
+import UserModel from "../db/models/UserModel.js"
 import auth from "../middlewares/auth.js"
 
 const postsRoute = ({ app }) => {
@@ -10,13 +11,19 @@ const postsRoute = ({ app }) => {
       session: { userId: user_id }
     } = req
 
-    const post = await PostModel.query().insertAndFetch({
-      title,
-      content,
-      is_published,
-      user_id
-    })
-    res.send(post)
+    const user = await UserModel.query().findById(user_id)
+
+    if (user && (user.userType === "admin" || user.userType === "autheur")) {
+      const post = await PostModel.query().insertAndFetch({
+        title,
+        content,
+        is_published,
+        user_id
+      })
+      res.send(post)
+    } else {
+      res.status(404).send({ error: "you can't update this post" })
+    }
   })
 
   //add comment
@@ -32,7 +39,7 @@ const postsRoute = ({ app }) => {
       post_id: postId,
       user_id: sessionUserId
     })
-    
+
 
     const post = await PostModel.query().findById(postId).where("posts.is_published", true)
 
@@ -48,7 +55,7 @@ const postsRoute = ({ app }) => {
       const comment = comments[i]
       const commentAuthor = await comment.$relatedQuery("author")
       const commentdate = new Date(comment.created_at)
-      postComments.push({id: comment.id, content: comment.content, created_at: commentdate.toLocaleDateString(), author: commentAuthor.displayName, user_id: commentAuthor.id})
+      postComments.push({ id: comment.id, content: comment.content, created_at: commentdate.toLocaleDateString(), author: commentAuthor.displayName, user_id: commentAuthor.id })
     }
     const author = await post.$relatedQuery("author")
     const date = new Date(post.publication_date)
@@ -66,8 +73,8 @@ const postsRoute = ({ app }) => {
       session: { userId: sessionUserId }
     } = req
 
-    const numberOfAffectedRows  = await CommentsModel.query().update({content: content}).where("id", commentId ).where("user_id", sessionUserId).where("post_id", postId)
-    
+    const numberOfAffectedRows = await CommentsModel.query().update({ content: content }).where("id", commentId).where("user_id", sessionUserId).where("post_id", postId)
+
 
     const post = await PostModel.query().findById(postId).where("posts.is_published", true)
 
@@ -83,7 +90,7 @@ const postsRoute = ({ app }) => {
       const comment = comments[i]
       const commentAuthor = await comment.$relatedQuery("author")
       const commentdate = new Date(comment.created_at)
-      postComments.push({id: comment.id, content: comment.content, created_at: commentdate.toLocaleDateString(), author: commentAuthor.displayName, user_id: commentAuthor.id})
+      postComments.push({ id: comment.id, content: comment.content, created_at: commentdate.toLocaleDateString(), author: commentAuthor.displayName, user_id: commentAuthor.id })
     }
     const author = await post.$relatedQuery("author")
     const date = new Date(post.publication_date)
@@ -121,6 +128,25 @@ const postsRoute = ({ app }) => {
     res.send({ posts: postsToSend, total: posts.total })
   })
 
+  //get post for editing
+  app.get("/posts/edit/:postId", async (req, res) => {
+    const {
+      params: { postId }
+    } = req
+
+    const post = await PostModel.query().findById(postId)
+
+    if (!post) {
+      res.status(404).send({ error: "post not found, or not published" })
+
+      return
+    }
+
+    res.send({
+      id: post.id, title: post.title, content: post.content, is_published: post.is_published
+    })
+  })
+
   //recuperer post from id
   app.get("/posts/:postId", async (req, res) => {
     const {
@@ -141,13 +167,13 @@ const postsRoute = ({ app }) => {
       const comment = comments[i]
       const commentAuthor = await comment.$relatedQuery("author")
       const commentdate = new Date(comment.created_at)
-      postComments.push({id: comment.id, content: comment.content, created_at: commentdate.toLocaleDateString(), author: commentAuthor.displayName, user_id: commentAuthor.id})
+      postComments.push({ id: comment.id, content: comment.content, created_at: commentdate.toLocaleDateString(), author: commentAuthor.displayName, user_id: commentAuthor.id })
     }
     const author = await post.$relatedQuery("author")
     const date = new Date(post.publication_date)
     res.send({
       id: post.id, title: post.title, content: post.content, publication_date: date.toLocaleDateString(), user_id: post.user_id,
-      author: author.displayName, comments: postComments
+      author: author.displayName, is_published: post.is_published, comments: postComments
     })
   })
 
@@ -179,9 +205,9 @@ const postsRoute = ({ app }) => {
     } = req
 
     const post = await PostModel.query()
-      .findById(postId)
+      .findById(postId).where("user_id", sessionUserId)
       .patch({
-        title: title, content: content, is_published: is_published, user_id: sessionUserId
+        title: title, content: content, is_published: is_published
       })
 
     if (!post) {
@@ -195,18 +221,35 @@ const postsRoute = ({ app }) => {
   app.delete("/posts/:postId", auth, async (req, res) => {
     const {
       params: { postId: pId },
+      session: { userId: sessionUserId }
     } = req
 
-    const post = await PostModel.query().findById(Number(pId))
+    const post = await PostModel.query().findById(Number(pId))//.where("user_id", sessionUserId)
 
     if (!post) {
       res.status(404).send({ error: "post not found" })
     } else {
-      //delet post from DB
-      await post.$query().delete()
-      //and delete post comments from DB
-      await post.$relatedQuery("comments").delete()
-      res.send({ message: "post deleted with success" })
+      let deleteDone = false
+
+      if (post.user_id === sessionUserId) {
+        await post.$relatedQuery("comments").delete()
+        await post.$query().delete()
+        deleteDone = true
+      } else {
+        const author = await UserModel.query().findById(sessionUserId)
+
+        if (author.userType === "admin") {
+          await post.$relatedQuery("comments").delete()
+          await post.$query().delete()
+          deleteDone = true
+        }
+      }
+
+      if (deleteDone) {
+        res.send({ message: "post deleted with success" })
+      } else {
+        res.status(404).send({ error: "you cannot delete this post" })
+      }
     }
   })
 
@@ -217,15 +260,37 @@ const postsRoute = ({ app }) => {
       session: { userId: sessionUserId }
     } = req
 
-    console.log("delete comments : " + cId)
-
-    const comment = await CommentsModel.query().findById(Number(cId)).where("comments.user_id", sessionUserId)
+    const comment = await CommentsModel.query().findById(Number(cId))
 
     if (!comment) {
       res.status(404).send({ error: "comment not found" })
     } else {
-      await comment.$query().delete()
-      res.send({ message: "comment deleted with success" })
+      let deleteDone = false
+
+      if (comment.user_id === sessionUserId) { //commentaire de la personne connecté
+        await comment.$query().delete()
+        deleteDone = true
+      } else {
+        const post = await comment.$relatedQuery("post")
+
+        if (post.user_id === sessionUserId) { //post de la personne connecté
+          await comment.$query().delete()
+          deleteDone = true
+        } else {
+          const author = await UserModel.query().findById(sessionUserId)
+
+          if (author.userType === "admin") { // la personne connecté est un admin
+            await comment.$query().delete()
+            deleteDone = true
+          }
+        }
+      }
+
+      if (deleteDone) {
+        res.send({ message: "comment deleted with success" })
+      } else {
+        res.status(404).send({ error: "you cannot delete this comment" })
+      }
     }
   })
 }
